fix bug in mixdown: double size check in bounds check w float arithmetic in body caused bounds violation due to loss of precision

+++ src/mixdown.c
@@ -147,7 +147,7 @@ float get_track_channel_chunk(Track *track, float *chunk, uint8_t channel, int32
 	    pthread_mutex_unlock(&cr->lock);
 	    continue;
 	}
-	double pos_in_clip_sframes = start_pos_sframes - cr->pos_sframes;
+	float pos_in_clip_sframes = start_pos_sframes - cr->pos_sframes;
 	int32_t end_pos = pos_in_clip_sframes + (step * len_sframes);
 	int32_t min, max;
 	if (end_pos >= pos_in_clip_sframes) {
@@ -171,7 +171,7 @@ float get_track_channel_chunk(Track *track, float *chunk, uint8_t channel, int32
 	    /* Clip overlaps */
 	    if (pos_in_clip_sframes >= 0 && pos_in_clip_sframes < cr_len - 1) { /* Truncate last sample to allow for interpolation */
 		float sample;
-		float clip_index_f = pos_in_clip_sframes + cr->in_mark_sframes;
+		float clip_index_f = pos_in_clip_sframes + (float)cr->in_mark_sframes;
 		if (fabs(step) != 1.0f) {
 		    int index_left = (int)floor(clip_index_f);
 		    float diff_left = clip_index_f - (float)index_left;
